Fix quantization parameters for concat result.
Note that this fix only applies to TF models containing
tf.FakeQuantWithMinMaxVars ops.

Currently, concat ops with fake quants as inputs don't get their outputs
properly quantized the `tfl-prepare-tf` pass. Later on in the
`tfl-prepare-quantize` pass , the concat result is quantized using the scale of
the first input. This generates incorrect inference results.

This change generates a new fake quant that consumes the output of the concat,
taking the min and max from across the fake quant inputs. As the result of the
concat is now properly quantized it avoids the aformentioned eccentric
behavior.

PiperOrigin-RevId: 665579759